Abstract out components from Home.js
